+++ source/include/Ostap/Bernstein.h
@@ -254,15 +254,15 @@ namespace Ostap
        *
        *  @param xmin low  edge for Bernstein polynomial
        *  @param xmax high edge for Bernstein polynomial
-       *  @param r  the list of real  roots of the polinomial
-       *  @param c  the list of complex roots (only one root from cc-pair is needed)
+       *  @param roots_real    the list of real  roots of the polinomial
+       *  @param roots_complex the list of complex roots (only one root from cc-pair is needed)
        */
       Bernstein 
       ( const double xmin , 
         const double xmax , 
-        const std::vector<double>&                 r = 
+        const std::vector<double>&                 roots_real  = 
         std::vector<double> () , 
-        const std::vector<std::complex<double> > & c = 
+        const std::vector<std::complex<double> > & roots_complex = 
         std::vector<std::complex<double> > ()  );
       // ======================================================================
       /** construct Bernstein polynomial from its roots
@@ -272,14 +272,14 @@ namespace Ostap
        *
        *  @param xmin low  edge for Bernstein polynomial
        *  @param xmax high edge for Bernstein polynomial
-       *  @param c  the list of complex roots (only one root from cc-pair is needed)
-       *  @param r  the list of real  roots of the polinomial
+       *  @param roots_complex the list of complex roots (only one root from cc-pair is needed)
+       *  @param roots_real    the list of real  roots of the polinomial
        */
       Bernstein 
       ( const double xmin , 
         const double xmax , 
-        const std::vector<std::complex<double> > & c ,
-        const std::vector<double>&                 r = 
+        const std::vector<std::complex<double> > & roots_complex ,
+        const std::vector<double>&                 roots_real    = 
         std::vector<double> () ) ;
       // ======================================================================
     public: // constructors from different plynomial types