+++ source/src/local_math.h
@@ -20,11 +20,11 @@ namespace
 {
   // ==========================================================================
   /// equality criteria for doubles
-  const Ostap::Math::Equal_To<double>            s_equal{} ; // equality criteria for doubles
+  const Ostap::Math::Equal_To<double>            s_equal{}      ; // equality criteria for doubles
   /// zero for doubles  
-  const Ostap::Math::Zero<double>                s_zero {} ; // zero for doubles
+  const Ostap::Math::Zero<double>                s_zero {}      ; // zero for doubles
   /// zero for vectors 
-  const Ostap::Math::Zero< std::vector<double> > s_vzero{} ; // zero for vectors
+  const Ostap::Math::Zero< std::vector<double> > s_vzero{}      ; // zero for vectors
   // ==========================================================================
   // Limits? 
   // ==========================================================================